Decouple stop_wait() call from pubnub_connection_cleanup(), call it 'manually' just before user cb
If we are to retry a connection, we must not call stop_wait() too soon;
some future paths in pubnub_handle_error() will not invoke it.

This makes the code annoyingly convoluted and spills out some of the
event handling internals to the user-facing functions.  I pondered
instead declaring that stop_wait() callback might not be definitive if
followed immediately by a timeout() or add_socket() call, but that makes
for even uglier public API.