Introduce "subscription channelset" to pubnub context
Subscribe/unsubscribe now not just starts/interrupts a subscription
HTTP request, but also modifies the subscription channelset; a subscribe
then works on the complete channelset.

This specifically means that a pattern like

	pubnub_subscribe(p, "chan1");
	...
	pubnub_subscribe(p, "chan2");

is not backwards-compatible now. Before, the second subscribe would
listen to just chan2, now it will subscribe to _both_ chan1 and chan2.