pubnub_finished_cb(): New finished_cb wrapper that cleans up beforehand
In some places, e.g. resubscribe_http_init(), we look at whether
p->finished_cb_data holds anything, and if there are leftovers from
a previous call (e.g. here_now), things blow up.  Therefore, we make
sure that at the point when we issue the callback, there are no
leftovers.

This fixes specifically a crash when we issue here_now and from
its callback perform a subscribe. here_now is one of the few functions
that have no callback wrapper that'd clean up finished_cb_data, and
the subscribe eventually descends to a resubscribe call, which looks
at finished_cb_data and from its existence it judges that we just
interrupted an ongoing subscribe.

So we now fulfill its implicit assumption that finished_cb_data is
always a NULL when there's no call in progress.