libpubnub-1.0: First stable release of the libpubnub API
This seems as good a moment as any, as our library seems to be gathering
non-trivial deployment.

We expect not to fix ourselves on the "1.0" version number overly; this
version indicator serves only to distinguish API/ABI compatibility and
we will bump it whenever we need to break something.  If we ever need
to make a truly major overhaul of the library, we'll probably call that
libpubnub2 or something.

In short, marking this as 1.0 is just a convenience for our users and
a stable reference point.

We also introduce PUBNUB_APIVER_MAJOR / PUBNUB_APIVER_MINOR #defines
in pubnub.h along this.